requirements.txt
Django>=6.0
django-jazzmin
# django-ses with SNS events support
django-ses[events]
# django-post-office for queued emails
django-post-office
boto3
